fixup! move newRepoRegistry() into it's own file
