+++ cmd/image-builder/list_images.go
@@ -2,31 +2,9 @@ package main
 
 import (
 	"io"
-	"os"
-	"strings"
 
 	"github.com/osbuild/images/pkg/distrofactory"
-	"github.com/osbuild/images/pkg/reporegistry"
 )
-
-// XXX: copied from "composer", should be exported there so
-// that we keep this in sync
-// XXX2: means we need to depend on osbuild-composer-common or something
-var repositoryConfigs = []string{
-	"/etc/osbuild-composer",
-	"/usr/share/osbuild-composer",
-}
-
-// XXX: move to pkg/reporegistry
-func newRepoRegistry() (*reporegistry.RepoRegistry, error) {
-	// useful for development/debugging, e.g. run:
-	// go build && IMAGE_BUILDER_EXTRA_REPOS_PATH=../../test/data ./image-builder
-	if extraReposPath := os.Getenv("IMAGE_BUILDER_EXTRA_REPOS_PATH"); extraReposPath != "" {
-		repositoryConfigs = append(repositoryConfigs, strings.Split(extraReposPath, ":")...)
-	}
-
-	return reporegistry.New(repositoryConfigs)
-}
 
 func getFilteredImages(filterExprs []string) ([]FilterResult, error) {
 	repos, err := newRepoRegistry()
